requirements.txt
tensorflow==2.18.0
pillow==11.0.0
numpy==1.26.4
psycopg2==2.9.10